Unify gaze buttons
